+++ zellij-server/src/ui/mod.rs
@@ -2,3 +2,4 @@ pub mod boundaries;
 pub mod overlay;
 pub mod pane_boundaries_frame;
 pub mod pane_contents_and_ui;
+pub mod loading_indication;
